Refactor helpers.py
