+++ jaxquantum/core/__init__.py
@@ -7,4 +7,4 @@ from .solvers import *
 from .qarray import *
 from .settings import SETTINGS
 from .dims import *
-from .helpers import *
+from .measurements import *
